Add web GUI for Password Generator and improve functionality
This commit adds a web GUI for the Password Generator application. The HTML file "web.html" contains the necessary elements for generating and displaying passwords. It also includes JavaScript functions to generate passwords and copy them to the clipboard. The password length is currently set to 12 characters, but it can be customized in the future. This update enhances the functionality of the application and provides a more user-friendly interface.